Model Loading In Progress

+++ Villain CES/Villain CES/ProjectileSystem.cpp
@@ -31,7 +31,7 @@ bool CProjectileSystem::CreateBulletProjectile(int indexofBullet,Clips* Gun)
 bool CProjectileSystem::Reload(Clips * Gun)
 {
 	if (Gun->nBulletsAvailables.size() != Gun->nSizeofClipl) {
-		for (int i = Gun->nBulletsAvailables.size(); i < Gun->nSizeofClipl; ++i) {
+		for (size_t i = Gun->nBulletsAvailables.size(); i < Gun->nSizeofClipl; ++i) {
 			Gun->nBulletsAvailables.push_back(true);
 		}
 		return true;
